requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary  # For PostgreSQL, or sqlite3 if testing locally
PyJWT            # For JWT token handling
emails           # For SMTP (optional, replace with sendgrid if preferred)
flask
flask-cors
pytest
requests
python-dotenv
bcrypt           # For password and PIN hashing
pyotp            # For TOTP MFA
slowapi          # For rate limiting